move tensor_to_textured_rect out of images.rs
